+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.5.0 - 2026-06-30
 
 ### Added
 - `github_delete_branch` to safely delete branches with merged-only checks against the default or provided base branch.